+++ shared/src/test/scala/edu/holycross/shot/ohco2/CorpusFilteringSpec.scala
@@ -192,4 +192,8 @@ urn:cts:greekLit:tlg5026.msA.hmt:1.5.comment#<div xmlns="http://www.tei-c.org/ns
     val strFromFilter = corpus.textContents(CtsUrn("urn:cts:greekLit:tlg5026.msA:1.3-1.4"))
     assert(strFromFilter.split("\n").size == 4)
   }
+
+  it should "correctly extract passages from multiple versions when twiddling on a notional work" in {
+    
+  }
 }